AgenticAI/langgraph/simple_chatbot/backend/requirements.txt
pydantic
langchain
langgraph
langchain-core
langchain-community
langchain-groq
python-dotenv
arxiv
wikipedia
fastapi
uvicorn
tavily-python
structlog
ujson
